[frontend] add search input in knowledge and files tabs
